+++ JOUR02/Job02/Aquatique.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 #include "Aquatique.h"
 
-Aquatique::Aquatique(double metre, double seconde) {
+Aquatique::Aquatique(int metre, int seconde) {
     this->metre = metre;
     this->seconde = seconde;
 }
 
-void Aquatique::swim(double speed) {
+void Aquatique::swim(int speed) {
     std::cout << "I'm swimming at " << speed << " m/s" << std::endl;
 }
