+++ JOUR02/Job02/Terrestre.cpp
@@ -2,12 +2,12 @@
 #include "Terrestre.h"
 
 
-Terrestre::Terrestre(double metre, double seconde) {
+Terrestre::Terrestre(int metre, int seconde) {
     this->metre = metre;
     this->seconde = seconde;
 }
 
-void Terrestre::walk(double speed) {
+void Terrestre::walk(int speed) {
     std::cout << "I'm walking at " << speed << " m/s" << std::endl;
 }
